@@ -5,7 +5,7 @@ author: westdabestdb <westdabestdb@gmail.com>
 homepage: https://github.com/westdabestdb/md2-tab-indicator
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
