@@ -9,17 +9,15 @@ enum MD2IndicatorSize {
 }
 
 class MD2Indicator extends Decoration {
-  final double indicatorHeight;
-  final Color indicatorColor;
+  final double? indicatorHeight;
+  final Color? indicatorColor;
   final MD2IndicatorSize indicatorSize;
 
   const MD2Indicator(
-      {@required this.indicatorHeight,
-      @required this.indicatorColor,
-      @required this.indicatorSize});
+      {this.indicatorHeight, this.indicatorColor, required this.indicatorSize});
 
   @override
-  _MD2Painter createBoxPainter([VoidCallback onChanged]) {
+  _MD2Painter createBoxPainter([VoidCallback? onChanged]) {
     return new _MD2Painter(this, onChanged);
   }
 }
@@ -27,27 +25,28 @@ class MD2Indicator extends Decoration {
 class _MD2Painter extends BoxPainter {
   final MD2Indicator decoration;
 
-  _MD2Painter(this.decoration, VoidCallback onChanged)
-      : assert(decoration != null),
-        super(onChanged);
+  _MD2Painter(this.decoration, VoidCallback? onChanged) : super(onChanged);
 
   @override
   void paint(Canvas canvas, Offset offset, ImageConfiguration configuration) {
-    assert(configuration != null);
-    assert(configuration.size != null);
-
-    Rect rect;
+    Rect? rect;
     if (decoration.indicatorSize == MD2IndicatorSize.full) {
-      rect = Offset(offset.dx,
-              (configuration.size.height - decoration.indicatorHeight ?? 3)) &
-          Size(configuration.size.width, decoration.indicatorHeight ?? 3);
+      rect = Offset(
+              offset.dx,
+              (configuration.size!.height -
+                  (decoration.indicatorHeight ?? 3))) &
+          Size(configuration.size!.width, decoration.indicatorHeight ?? 3);
     } else if (decoration.indicatorSize == MD2IndicatorSize.normal) {
-      rect = Offset(offset.dx + 6,
-              (configuration.size.height - decoration.indicatorHeight ?? 3)) &
-          Size(configuration.size.width - 12, decoration.indicatorHeight ?? 3);
+      rect = Offset(
+              offset.dx + 6,
+              (configuration.size!.height -
+                  (decoration.indicatorHeight ?? 3))) &
+          Size(configuration.size!.width - 12, decoration.indicatorHeight ?? 3);
     } else if (decoration.indicatorSize == MD2IndicatorSize.tiny) {
-      rect = Offset(offset.dx + configuration.size.width / 2 - 8,
-              (configuration.size.height - decoration.indicatorHeight ?? 3)) &
+      rect = Offset(
+              offset.dx + configuration.size!.width / 2 - 8,
+              (configuration.size!.height -
+                  (decoration.indicatorHeight ?? 3))) &
           Size(16, decoration.indicatorHeight ?? 3);
     }
 
@@ -55,7 +54,7 @@ class _MD2Painter extends BoxPainter {
     paint.color = decoration.indicatorColor ?? Color(0xff1967d2);
     paint.style = PaintingStyle.fill;
     canvas.drawRRect(
-        RRect.fromRectAndCorners(rect,
+        RRect.fromRectAndCorners(rect!,
             topRight: Radius.circular(8), topLeft: Radius.circular(8)),
         paint);
   }
